(PC-26529)[PRO] refactor: reorganize main redux files
