+++ pro/src/components/ImageUploader/ButtonAppPreview/__specs__/ButtonAppPreview.spec.tsx
@@ -3,7 +3,7 @@ import { userEvent } from '@testing-library/user-event'
 import React from 'react'
 
 import { UploaderModeEnum } from 'components/ImageUploader/types'
-import { RootState } from 'store/reducers'
+import { RootState } from 'store/rootReducer'
 
 import ButtonAppPreview, { ButtonAppPreviewProps } from '../ButtonAppPreview'
