+++ pro/src/components/ImageUploader/ButtonImageDelete/__specs__/ButtonImageDelete.spec.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
 import React from 'react'
 
-import { RootState } from 'store/reducers'
+import { RootState } from 'store/rootReducer'
 
 import ButtonImageDelete, { ButtonImageDeleteProps } from '../ButtonImageDelete'
