+++ pro/src/hooks/useActiveFeature.ts
@@ -1,7 +1,7 @@
 import { useSelector } from 'react-redux'
 
 import { isFeatureActive } from 'store/features/selectors'
-import { RootState } from 'store/reducers'
+import { RootState } from 'store/rootReducer'
 
 const useActiveFeature = (featureName: string): boolean => {
   const isActive = useSelector((state: RootState) =>
