+++ pro/src/pages/IndividualOfferWizard/Confirmation/__specs__/Confirmation.spec.tsx
@@ -11,7 +11,7 @@ import {
   IndividualOfferContext,
 } from 'context/IndividualOfferContext'
 import { IndividualOffer } from 'core/Offers/types'
-import { RootState } from 'store/reducers'
+import { RootState } from 'store/rootReducer'
 import { offerVenueFactory } from 'utils/apiFactories'
 import {
   individualOfferContextFactory,
