+++ pro/src/screens/IndividualOffer/ActionBar/ActionBar.tsx
@@ -10,7 +10,7 @@ import useNotification from 'hooks/useNotification'
 import fullLeftIcon from 'icons/full-left.svg'
 import fullRightIcon from 'icons/full-right.svg'
 import fullValidateIcon from 'icons/full-validate.svg'
-import { RootState } from 'store/reducers'
+import { RootState } from 'store/rootReducer'
 import { Button, ButtonLink, SubmitButton } from 'ui-kit'
 import { ButtonVariant, IconPositionEnum } from 'ui-kit/Button/types'
 import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
