+++ pro/src/screens/IndividualOffer/StocksThing/__specs__/StocksThing.edition.spec.tsx
@@ -21,7 +21,7 @@ import {
   getIndividualOfferUrl,
 } from 'core/Offers/utils/getIndividualOfferUrl'
 import Stocks from 'pages/IndividualOfferWizard/Stocks/Stocks'
-import { RootState } from 'store/reducers'
+import { RootState } from 'store/rootReducer'
 import { individualGetOfferStockResponseModelFactory } from 'utils/individualApiFactories'
 import { renderWithProviders } from 'utils/renderWithProviders'
