+++ pro/src/screens/OfferEducational/__specs__/OfferEducationalCreationAccessibilityStep.spec.tsx
@@ -2,7 +2,7 @@ import { screen, waitFor } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
 import React from 'react'
 
-import { RootState } from 'store/reducers'
+import { RootState } from 'store/rootReducer'
 import { renderWithProviders } from 'utils/renderWithProviders'
 
 import OfferEducational from '../'
