+++ pro/src/store/features/selectors.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit'
 
-import { RootState } from 'store/reducers'
+import { RootState } from 'store/rootReducer'
 
 export const isFeatureActive = (
   state: RootState,
