+++ pro/src/store/notifications/selectors.ts
@@ -1,5 +1,5 @@
 import { Notification } from 'store/notifications/reducer'
-import { RootState } from 'store/reducers'
+import { RootState } from 'store/rootReducer'
 
 export const notificationSelector = (state: RootState): Notification | null => {
   return state.notification.notification
