+++ pro/src/store/testUtils.ts
@@ -1,11 +1,11 @@
 import { FeatureResponseModel, UserRole } from 'apiClient/v1'
-import createStore from 'store'
 import { initialState as featuresInitialState } from 'store/features/reducer'
 import { initialState as notificationInitialState } from 'store/notifications/reducer'
 import { initialState as offersInitialState } from 'store/offers/reducer'
+import createStore from 'store/store'
 import { initialState as userInitialState } from 'store/user/reducer'
 
-import { RootState } from './reducers'
+import { RootState } from './rootReducer'
 
 export const configureTestStore = (overrideData?: Partial<RootState>) => {
   const initialData = {
