+++ pro/src/store/user/selectors.ts
@@ -1,3 +1,3 @@
-import { RootState } from 'store/reducers'
+import { RootState } from 'store/rootReducer'
 
 export const selectCurrentUser = (state: RootState) => state.user.currentUser
